+++ src/SignalProcessing/StatFilter.cpp
@@ -408,7 +408,7 @@ static int oldtarget, oldoutcome;
      {
         sample= 0;
 
-        for(in_channel=0; in_channel<SignalE->Channels(); in_channel++)
+        for(in_channel=0; in_channel<(int)SignalE->Channels(); in_channel++)
          {
          value= SignalE->GetValue(in_channel, sample);
